move ResourceActionsMenu component to its own file
