+++ ui/src/app/components/resource-card.tsx
@@ -1,13 +1,10 @@
 import * as fp from 'lodash/fp';
 import * as React from 'react';
-
 import {CSSProperties, PropsWithChildren} from 'react';
 
-import {Clickable, MenuItem} from 'app/components/buttons';
+import {Clickable} from 'app/components/buttons';
 import {ResourceCardBase} from 'app/components/card';
 import {FlexColumn, FlexRow} from 'app/components/flex';
-import {SnowmanIcon} from 'app/components/icons';
-import {PopupTrigger, TooltipTrigger} from 'app/components/popups';
 import colors from 'app/styles/colors';
 import {formatWorkspaceResourceDisplayDate, reactStyles} from 'app/utils';
 import {AnalyticsTracker} from 'app/utils/analytics';
@@ -24,6 +21,7 @@ import {
   isNotebook,
 } from 'app/utils/resources';
 import {WorkspaceResource} from 'generated/fetch';
+import {Action, ResourceActionsMenu} from './resource-actions-menu';
 
 const styles = reactStyles({
   card: {
@@ -69,42 +67,6 @@ const styles = reactStyles({
     flexDirection: 'column'
   }
 });
-
-interface Action {
-  icon: string;
-  displayName: string;
-  onClick: () => void;
-  disabled: boolean;
-  hoverText?: string;
-}
-
-const ResourceActionsMenu = (props: {actions: Action[]}) => {
-  const {actions} = props;
-  return <PopupTrigger
-      data-test-id='resource-card-menu'
-      side='bottom'
-      closeOnClick
-      content={
-        <React.Fragment>
-          {actions.map((action, i) => {
-            return (
-                <TooltipTrigger key={i} content={action.hoverText}>
-                  <MenuItem
-                      icon={action.icon}
-                      onClick={() => action.onClick()}
-                      disabled={action.disabled}>
-                    {action.displayName}
-                  </MenuItem>
-                </TooltipTrigger>);
-          })}
-        </React.Fragment>
-      }
-  >
-    <Clickable data-test-id='resource-menu'>
-      <SnowmanIcon disabled={false}/>
-    </Clickable>
-  </PopupTrigger>;
-};
 
 const StyledResourceType = (props: {resource: WorkspaceResource}) => {
   const {resource} = props;
@@ -168,7 +130,6 @@ const ResourceNavigation = (props: NavProps) => {
 interface Props {
   actions: Action[];
   resource: WorkspaceResource;
-  menuOnly: boolean;  // use this component strictly for its actions, without rendering the card
 }
 
 class ResourceCard extends React.Component<Props, {}> {
@@ -178,9 +139,8 @@ class ResourceCard extends React.Component<Props, {}> {
   }
 
   render() {
-    const {resource, menuOnly} = this.props;
-    return menuOnly ? <ResourceActionsMenu actions={this.props.actions}/> :
-        <ResourceCardBase style={styles.card}
+    const {resource} = this.props;
+    return <ResourceCardBase style={styles.card}
                           data-test-id='card'>
           <FlexColumn style={{alignItems: 'flex-start'}}>
             <FlexRow style={{alignItems: 'flex-start'}}>
@@ -199,7 +159,6 @@ class ResourceCard extends React.Component<Props, {}> {
 }
 
 export {
-  Action,
   ResourceCard,
   ResourceNavigation,
   StyledResourceType,
