+++ ui/src/app/pages/data/cohort-review/cohort-review-resource-card.tsx
@@ -2,7 +2,8 @@ import * as fp from 'lodash/fp';
 import * as React from 'react';
 
 import {RenameModal} from 'app/components/rename-modal';
-import {Action, canDelete, canWrite, ResourceCard} from 'app/components/resource-card';
+import {Action, ResourceActionsMenu} from 'app/components/resource-actions-menu';
+import {canDelete, canWrite, ResourceCard} from 'app/components/resource-card';
 import {withConfirmDeleteModal, WithConfirmDeleteModalProps} from 'app/components/with-confirm-delete-modal';
 import {withErrorModal, WithErrorModalProps} from 'app/components/with-error-modal';
 import {withSpinnerOverlay, WithSpinnerOverlayProps} from 'app/components/with-spinner-overlay';
@@ -97,11 +98,11 @@ export const CohortReviewResourceCard = fp.flow(
                    oldName={getDisplayName(resource)}
                    existingNames={this.props.existingNameList}/>
       }
-      <ResourceCard
-          resource={resource}
-          actions={this.actions}
-          menuOnly={menuOnly}
-      />
+      {menuOnly ? <ResourceActionsMenu actions={this.actions}/> :
+          <ResourceCard
+              resource={resource}
+              actions={this.actions}
+          />}
     </React.Fragment>;
   }
 });
