+++ ui/src/app/pages/data/concept/concept-set-resource-card.tsx
@@ -4,7 +4,8 @@ import * as React from 'react';
 import {CopyModal} from 'app/components/copy-modal';
 import {DataSetReferenceModal} from 'app/components/data-set-reference-modal';
 import {RenameModal} from 'app/components/rename-modal';
-import {Action, canDelete, canWrite, ResourceCard} from 'app/components/resource-card';
+import {Action, ResourceActionsMenu} from 'app/components/resource-actions-menu';
+import {canDelete, canWrite, ResourceCard} from 'app/components/resource-card';
 import {withConfirmDeleteModal, WithConfirmDeleteModalProps} from 'app/components/with-confirm-delete-modal';
 import {withErrorModal, WithErrorModalProps} from 'app/components/with-error-modal';
 import {withSpinnerOverlay, WithSpinnerOverlayProps} from 'app/components/with-spinner-overlay';
@@ -157,11 +158,11 @@ export const ConceptSetResourceCard = fp.flow(
             return this.deleteConceptSet();
           }}/>
       }
-      <ResourceCard
+      {menuOnly ? <ResourceActionsMenu actions={this.actions}/> :
+          <ResourceCard
           resource={resource}
           actions={this.actions}
-          menuOnly={menuOnly}
-      />
+      />}
     </React.Fragment>;
   }
 });
